- Fix small datatype warning in release
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@1848 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/Map.cpp
@@ -64,7 +64,7 @@ void Map::addPlayer(Player *player) {
 	if (player->getActiveBuffs()->getActiveSkillLevel(9101004) == 0)
 		MapPacket::showPlayer(player);
 	if (timer > 0)
-		MapPacket::showTimer(player, timer - (time(0) - timerstart));
+		MapPacket::showTimer(player, timer - static_cast<int32_t>(time(0) - timerstart));
 }
 
 void Map::removePlayer(Player *player) {
